vmlinux: unify x86_64 and arm64 vmlinux CO-RE header files
Closes: #1133
