+++ tracee-ebpf/Makefile
@@ -135,7 +135,8 @@ $(OUT_BPF_CORE): $(BPF_SRC) $(LIBBPF_HEADERS) $(CMD_CLANG)
 	$(CMD_CLANG) \
 		-I $(LIBBPF_HEADERS)/bpf \
 		-I $(BPF_HEADERS) \
-		-I tracee/$(ARCH) \
+		-I tracee/vmlinux \
+		-I tracee/vmlinux/$(ARCH) \
 		-DCORE \
 		-D__TARGET_ARCH_$(linux_arch) \
 		-D__BPF_TRACING__ \
