--- tracee-ebpf/tracee/x86_64/vmlinux.h
@@ -1,1 +1,0 @@
-vmlinux-core.h
